+++ app/controllers/journeys_controller.rb
@@ -23,6 +23,15 @@ class JourneysController < ApplicationController
     @specification_template = Liquid::Template.parse(
       @journey.liquid_template, error_mode: :strict
     )
+
+    specification_html = @specification_template.render(@answers)
+
+    respond_to do |format|
+      format.html
+      format.docx do
+        render docx: "specification.docx", content: specification_html, layout: "specficiation"
+      end
+    end
   end
 
   private